Fix resizing crash due to camera close timing
Sometimes the camera would not close fast enough on resize events, when
opening the camera after the surface was re-created, it could get stuck
in an endless loop due to a sleep call. This removes that call and uses
postDelayed to not block the camera thread.

Also clean up a memory leak with AHBManager calling the destructor
instead of delete for it's members.

Also, moved AHBManger logic out of ImageReaderListener to ensure that it
is created after VulkanInstance creation and destroyed at the right
time.